--- k8s/charts/rtsp-camera-simulator/Chart.yaml
@@ -1,14 +1,0 @@
-apiVersion: v2
-name: rtsp-camera-simulator
-description: RTSP camera simulator
-type: application
-
-# This is the chart version. This version number should be incremented each time you make changes
-# to the chart and its templates, including the app version.
-# Versions are expected to follow Semantic Versioning (https://semver.org/)
-version: 0.1.0
-
-# This is the version number of the application being deployed. This version number should be
-# incremented each time you make changes to the application. Versions are not expected to
-# follow Semantic Versioning. They should reflect the version the application is using.
-appVersion: 0.7.0
