--- k8s/charts/rtsp-camera-simulator/templates/deployment.yaml
@@ -1,41 +1,0 @@
-#
-# Copyright (c) Dell Inc., or its subsidiaries. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: {{ $.Release.Name }}
-  labels: 
-    release: {{ $.Release.Name }}
-spec:
-  selector:
-    matchLabels:
-      release: {{ $.Release.Name }}
-  template:
-    metadata:
-      labels:
-        release: {{ $.Release.Name }}
-    spec:
-      containers:
-      - name: rtsp-camera-simulator
-        image: {{ .Values.image.repository | quote }}
-        imagePullPolicy: {{ .Values.image.pullPolicy | quote }}
-        command: ["rtsp-camera-simulator"]
-        ports:
-        - name: rtsp
-          containerPort: {{ .Values.rtspPort }}
-          protocol: TCP
-        env:
-        - name: CAMERA_PORT
-          value: {{ .Values.rtspPort | quote }}
-        {{- range $key, $value := .Values.appParameters }}
-        - name: {{ $key | quote }}
-          value: {{ $value | quote }}
-        {{- end }}
